+++ Sources/AlloyUIKitControls/BufferingIndicator.swift
@@ -1,6 +1,6 @@
 //
 //  BufferingIndicator.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //