+++ Sources/AlloyUIKitControls/CustomStatusBar.swift
@@ -1,6 +1,6 @@
 //
 //  CustomStatusBar.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //