+++ Sources/AlloyUIKitControls/DefaultControlOverlay.swift
@@ -1,6 +1,6 @@
 //
 //  DefaultControlOverlay.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //