+++ Sources/AlloyUIKitControls/LandscapeControlPanel.swift
@@ -1,6 +1,6 @@
 //
 //  LandscapeControlPanel.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //