+++ Sources/AlloyUIKitControls/NetworkSpeedMonitor.swift
@@ -1,6 +1,6 @@
 //
 //  NetworkSpeedMonitor.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //