+++ Sources/AlloyUIKitControls/ProgressSlider.swift
@@ -1,6 +1,6 @@
 //
 //  ProgressSlider.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //