+++ Sources/AlloyUIKitControls/VolumeAndBrightnessHUD.swift
@@ -1,6 +1,6 @@
 //
 //  VolumeAndBrightnessHUD.swift
-//  AlloyControlView
+//  AlloyUIKitControls
 //
 //  Created by Sun on 2026/4/14.
 //